+++ audbackend/core/backend.py
@@ -1,12 +1,9 @@
 import errno
-import glob
 import os
 import re
-import shutil
 import tempfile
 import typing
 
-import audfactory
 import audeer
 
 from audbackend.core import utils
@@ -282,7 +279,8 @@ class Backend:
             ValueError: if ``path`` does not end on file extension
 
         Example:
-            >>> backend = FileSystem('~/my-host', 'data')
+            >>> import audbackend
+            >>> backend = audbackend.FileSystem('~/my-host', 'data')
             >>> path = backend.path('media/archive1.zip', '1.0.0')
             >>> os.path.basename(path)
             'archive1-1.0.0.zip'
@@ -503,287 +501,3 @@ class Backend:
         name = audeer.basename_wo_ext(file, ext=ext)
         vs = self._versions(folder, name)
         return audeer.sort_versions(vs)
-
-
-class Artifactory(Backend):
-    r"""Artifactory backend.
-
-    Stores files and archives on Artifactory.
-
-    Args:
-        host: host address
-        repository: repository name
-
-    """
-
-    def __init__(
-            self,
-            host,
-            repository,
-    ):
-        super().__init__(host, repository)
-
-    def _checksum(
-            self,
-            path: str,
-    ) -> str:
-        r"""MD5 checksum of file on backend."""
-        return audfactory.checksum(path)
-
-    def _path(
-            self,
-            folder: str,
-            name: str,
-            ext: str,
-            version: str,
-    ) -> str:
-        r"""File path on backend."""
-        server_url = audfactory.url(
-            self.host,
-            repository=self.repository,
-            group_id=audfactory.path_to_group_id(folder),
-            name=name,
-            version=version,
-        )
-        return f'{server_url}/{name}-{version}{ext}'
-
-    def _exists(
-            self,
-            path: str,
-    ) -> bool:
-        r"""Check if file exists on backend."""
-        try:
-            # Can lead to RuntimeError: 404 page not found
-            return audfactory.path(path).exists()
-        except RuntimeError:  # pragma: nocover
-            return False
-
-    def _get_file(
-            self,
-            src_path: str,
-            dst_path: str,
-    ):
-        r"""Get file from backend."""
-        audfactory.download(src_path, dst_path, verbose=False)
-
-    def _glob(
-            self,
-            pattern: str,
-            folder: typing.Optional[str],
-    ) -> typing.List[str]:
-        r"""Return matching files names."""
-        if folder is not None:
-            group_id = audfactory.path_to_group_id(folder)
-        else:
-            group_id = None
-        url = audfactory.url(
-            self.host,
-            repository=self.repository,
-            group_id=group_id,
-        )
-        path = audfactory.path(url)
-        try:
-            result = [str(x) for x in path.glob(pattern)]
-        except RuntimeError:  # pragma: no cover
-            result = []
-        return result
-
-    def _put_file(
-            self,
-            src_path: str,
-            dst_path: str,
-    ):
-        r"""Put file to backend."""
-        audfactory.deploy(src_path, dst_path)
-
-    def _remove_file(
-            self,
-            path: str,
-    ):
-        r"""Remove file from backend."""
-        audfactory.path(path).unlink()
-
-    def _versions(
-            self,
-            folder: str,
-            name: str,
-    ) -> typing.List[str]:
-        r"""Versions of a file."""
-        group_id = audfactory.path_to_group_id(folder)
-        return audfactory.versions(self.host, self.repository, group_id, name)
-
-
-class FileSystem(Backend):
-    r"""File system backend.
-
-    Stores files and archives on a file system.
-
-    Args:
-        host: host directory
-        repository: repository name
-
-    """
-    def __init__(
-            self,
-            host: str,
-            repository: str,
-    ):
-        super().__init__(audeer.safe_path(host), repository)
-
-    def _checksum(
-            self,
-            path: str,
-    ) -> str:
-        r"""MD5 checksum of file on backend."""
-        return utils.md5(path)
-
-    def _path(
-            self,
-            folder: str,
-            name: str,
-            ext: str,
-            version: str,
-    ) -> str:
-        r"""File path on backend."""
-        path = os.path.join(
-            self.host,
-            self.repository,
-            folder.replace(self.sep, os.path.sep),
-            name,
-        )
-        if version is not None:
-            path = os.path.join(
-                path,
-                version,
-                f'{name}-{version}{ext}',
-            )
-        return path
-
-    def _exists(
-            self,
-            path: str,
-    ) -> bool:
-        r"""Check if file exists on backend."""
-        return os.path.exists(path)
-
-    def _get_file(
-            self,
-            src_path: str,
-            dst_path: str,
-    ):
-        r"""Get file from backend."""
-        shutil.copy(src_path, dst_path)
-
-    def _glob(
-            self,
-            pattern: str,
-            folder: typing.Optional[str],
-    ) -> typing.List[str]:
-        r"""Return matching files names."""
-        if folder is None:
-            folder = ''
-        pattern = pattern.replace(self.sep, os.path.sep)
-        folder = folder.replace(self.sep, os.path.sep)
-        root = os.path.join(self.host, self.repository)
-        path = os.path.join(root, folder, pattern)
-        matches = glob.glob(path, recursive=True)
-        return [os.path.join(root, folder, match) for match in matches]
-
-    def _put_file(
-            self,
-            src_path: str,
-            dst_path: str,
-    ):
-        r"""Put file to backend."""
-        audeer.mkdir(os.path.dirname(dst_path))
-        shutil.copy(src_path, dst_path)
-
-    def _remove_file(
-            self,
-            path: str,
-    ):
-        r"""Remove file from backend."""
-        os.remove(path)
-
-    def _versions(
-            self,
-            folder: str,
-            name: str,
-    ) -> typing.List[str]:
-        r"""Versions of a file."""
-        root = self._path(folder, name, '', None)
-        vs = []
-        if os.path.exists(root):
-            vs = [
-                v for v in os.listdir(root)
-                if os.path.isdir(os.path.join(root, v))
-            ]
-        return vs
-
-
-backends = {}
-r"""Backend cache."""
-
-backend_registry = {
-    'file-system': FileSystem,
-    'artifactory': Artifactory,
-}
-r"""Backend registry."""
-
-
-def create(
-        name: str,
-        host: str,
-        repository: str,
-) -> Backend:
-    r"""Create backend.
-
-    Args:
-        name: backend registry name
-        host: host address
-        repository: repository name
-
-    Returns:
-        backend object
-
-    Raises:
-        ValueError: if registry name does not exist
-
-    """
-    if name not in backend_registry:
-        raise ValueError(
-            'A backend class with name '
-            f"'{name} "
-            'does not exist.'
-            "Use 'register_backend()' to register one."
-        )
-
-    if name not in backends:
-        backends[name] = {}
-    if host not in backends[name]:
-        backends[name][host] = {}
-    if repository not in backends[name][host]:
-        backend = backend_registry[name]
-        backends[name][host][repository] = backend(host, repository)
-    return backends[name][host][repository]
-
-
-def register(
-        name: str,
-        cls: typing.Type[Backend],
-):
-    r"""Register backend.
-
-    If a backend with this name already exists,
-    it will be overwritten.
-
-    Args:
-        name: backend registry name
-        cls: backend class
-
-    """
-    backend_registry[name] = cls
-
-
-register('artifactory', Artifactory)
-register('file-system', FileSystem)
